pyproject.toml
[project]
name = "hrm"
version = "0.1.0"
description = ""
authors = [
    {name = "NandaKishore1601",email = "nandhakish2001@gmail.com"}
]
readme = "README.md"
requires-python =  ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.14,<0.4.0)",
    "fastapi[standard]",
    "pydantic",
    "python-dotenv",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "openai (>=1.59.8,<2.0.0)",
    "langchain-openai (>=0.3.0,<0.4.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "asyncmy (>=0.2.10,<0.3.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "databases (>=0.9.0,<0.10.0)",
    "python-docx (>=1.1.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
